--- php_modules/plugins/page/viewmodels/AdminTemplates.php
@@ -1,156 +1,0 @@
-<?php
-
-/**
- * SPT software - ViewModel
- * 
- * @project: https://github.com/smpleader/spt
- * @author: Pham Minh - smpleader
- * @description: A simple View Model
- * 
- */
-
-namespace App\plugins\page\viewmodels;
-
-use SPT\Web\Gui\Form;
-use SPT\Web\Gui\Listing;
-use SPT\Web\ViewModel;
-
-class AdminTemplates extends ViewModel
-{
-    public static function register()
-    {
-        return [
-            'layout'=>[
-                'backend.template.list',
-                'backend.template.list.row',
-                'backend.template.list.filter'
-            ]
-        ];
-    }
-
-    public function list()
-    {
-        $filter = $this->filter()['form'];
-        $limit  = $filter->getField('limit')->value;
-        $sort   = $filter->getField('sort')->value;
-        $search = trim($filter->getField('search')->value);
-        $page = $this->state('page', 1, 'int', 'get', 'template.page');
-        if ($page <= 0) $page = 1;
-
-        $where = ['status <> 2'];
-
-        if (!empty($search)) {
-            $where[] = "(`title` LIKE '%" . $search . "%') OR (`path` LIKE '%" . $search . "%')";
-        }
-
-        $start  = ($page - 1) * $limit;
-        $sort = $sort ? $sort : 'title asc';
-
-        $result = $this->TemplateEntity->list($start, $limit, $where, $sort);
-        $total = $this->TemplateEntity->getListTotal();
-        
-        if (!$result) {
-            $result = [];
-            $total = 0;
-        }
-
-        $limit = $limit == 0 ? $total : $limit;
-        $list   = new Listing($result, $total, $limit, $this->getColumns());
-        return [
-            'list' => $list,
-            'page' => $page,
-            'start' => $start,
-            'sort' => $sort,
-            'user_id' => $this->user->get('id'),
-            'url' => $this->router->url(),
-            'link_list' => $this->router->url('templates'),
-            'title_page' => 'Templates',
-            'link_form' => $this->router->url('template'),
-            'token' => $this->token->value(),
-        ];
-    }
-
-    public function getColumns()
-    {
-        return [
-            'num' => '#',
-            'name' => 'Name',
-            'description' => 'Description',
-            'parent' => 'Parent',
-            'col_last' => ' ',
-        ];
-    }
-
-    protected $_filter;
-    public function filter()
-    {
-        if (null === $this->_filter) :
-            $data = [
-                'search' => $this->state('search', '', '', 'post', 'template.search'),
-                'limit' => $this->state('limit', 10, 'int', 'post', 'template.limit'),
-                'sort' => $this->state('sort', '', '', 'post', 'template.sort')
-            ];
-            $filter = new Form($this->getFilterFields(), $data);
-
-            $this->_filter = $filter;
-        endif;
-
-        return ['form' => $this->_filter];
-    }
-
-    public function getFilterFields()
-    {
-        return [
-            'search' => [
-                'text',
-                'default' => '',
-                'showLabel' => false,
-                'formClass' => 'form-control h-full input_common w_full_475',
-                'placeholder' => 'Search..'
-            ],
-            'status' => [
-                'option',
-                'default' => '1',
-                'formClass' => 'form-select',
-                'options' => [
-                    ['text' => '--', 'value' => ''],
-                    ['text' => 'Show', 'value' => '1'],
-                    ['text' => 'Hide', 'value' => '0'],
-                ],
-                'showLabel' => false
-            ],
-            'limit' => [
-                'option',
-                'formClass' => 'form-select',
-                'default' => 20,
-                'options' => [
-                    ['text' => '20', 'value' => 20],
-                    ['text' => '50', 'value' => 50],
-                    ['text' => 'All', 'value' => 0],
-                ],
-                'showLabel' => false
-            ],
-            'sort' => [
-                'option',
-                'formClass' => 'form-select',
-                'default' => 'name asc',
-                'options' => [
-                    ['text' => 'title ascending', 'value' => 'title asc'],
-                    ['text' => 'title descending', 'value' => 'title desc'],
-                ],
-                'showLabel' => false
-            ]
-        ];
-    }
-
-    public function row($layoutData, $viewData)
-    {
-        $row = $viewData['list']->getRow();
-        return [
-            'item' => $row,
-            'index' => $viewData['list']->getIndex(),
-        ];
-    }
-
-
-}
